Add useRegister hook
