+++ src/app/tnlcm/register/Register.module.css
@@ -1,10 +1,12 @@
 .register-container {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
     display: flex;
     flex-direction: column;
-    align-items: center;
     justify-content: center;
-    height: 100vh;
-    background-color: #f0f0f0;
+    align-items: center;
 }
   
 .register-form {
