+++ src/app/tnlcm/register/page.js
@@ -1,37 +1,23 @@
 "use client"
 
-import { useState } from "react";
-import { useRouter } from "next/navigation";
+import useRegister from "@/hooks/useRegister";
 import Button from "@/components/elements/Button";
 import Input from "@/components/elements/Input";
-import { registerUser } from "@/lib/apiHandler";
 import styles from "./Register.module.css";
 
 export default function RegisterPage() {
-  const [username, setUsername] = useState("");
-  const [password, setPassword] = useState("");
-  const [email, setEmail] = useState("");
-  const [org, setOrg] = useState("");
-  const router = useRouter();
-
-  const handleRegister = async (e) => {
-    e.preventDefault();
-    if (!username || !password || !email || !org) {
-      alert("All fields are required");
-    }
-    try {
-      await registerUser(username, password, email, org);
-      router.push("/tnlcm/login");
-    } catch (error) {
-      alert(error);
-    }
-  };
-
-  const handleKeyPress = (e) => {
-    if (e.key === "Enter") {
-      handleRegister(e);
-    }
-  };
+  const {
+    username,
+    setUsername,
+    password,
+    setPassword,
+    email,
+    setEmail,
+    org,
+    setOrg,
+    handleRegister,
+    handleKeyPress
+  } = useRegister();
 
   return (
     <div className={styles["register-container"]}>
